refactor: add createIssue function
